feat: getStoreDetail api 연동
