+++ frontend/src/interface/index.ts
@@ -4,3 +4,4 @@ export * from "./customer/signup";
 export * from "./common/layout";
 export * from "./seller/signup";
 export * from "./seller/product";
+export * from "./seller/base";
