+++ frontend/src/interface/seller/signup.ts
@@ -1,5 +1,13 @@
 import type { JSX } from "react";
 import type { ProductRequestType } from "./product";
+import type {
+  AddressInfoType,
+  ImageInfoType,
+  OperationTimeType,
+  PaymentInfoType,
+  SnsInfoType,
+  StoreInfoType,
+} from "./base";
 
 export interface SellerSignupProps {
   pageIdx: number;
@@ -14,45 +22,6 @@ export interface SellerSignupPkgProps extends SellerSignupProps {
 export type PageComponent =
   | ((props: SellerSignupProps) => JSX.Element)
   | ((props: SellerSignupPkgProps) => JSX.Element);
-
-export type StoreInfoType = {
-  store_name: string;
-  store_introduction: string;
-  store_phone: string;
-};
-
-export type SnsInfoType = {
-  instagram?: string;
-  facebook?: string;
-  x?: string;
-  homepage?: string;
-};
-
-export type AddressInfoType = {
-  postal_code: string;
-  address: string;
-  detail_address: string;
-  sido: string;
-  sigungu: string;
-  bname: string;
-  lat: string;
-  lng: string;
-};
-
-export type OperationTimeType = {
-  day_of_week: number;
-  open_time: string;
-  pickup_start_time: string;
-  pickup_end_time: string;
-  close_time: string;
-  is_open_enabled: boolean;
-};
-
-export type PaymentInfoType = {
-  portone_store_id: string;
-  portone_channel_id: string;
-  portone_secret_key: string;
-};
 
 export type SignupRequestType = StoreInfoType & {
   sns_info: SnsInfoType;
@@ -76,16 +45,9 @@ export type SignupImageRequestType = {
   images: ImagesType[];
 };
 
-export type SignupImageBaseType = {
-  image_id: string;
-  image_url: string;
-  is_main: boolean;
-  display_order: number;
-};
-
 export type SignupImageResponseType = {
   store_id: string;
-  images: SignupImageBaseType[];
+  images: ImageInfoType[];
   total: number;
 };
